requirements.txt
PySide6
paddleocr==3.0.0
paddlepaddle==3.0.0
setuptools
Pillow
pdf2image
rapidfuzz
flask
flask-cors